ai-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
youtube-transcript-api==0.6.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
hypothesis==6.88.1